requirements.txt
# ============ Core Framework ============
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10

# ============ Vector Store ============
chromadb>=0.4.22

# ============ Embeddings ============
sentence-transformers>=2.2.2
huggingface-hub>=0.20.0

# ============ Local LLM ============
llama-cpp-python>=0.2.0

# ============ PDF Processing ============
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0

# ============ OCR (optional) ============
pytesseract>=0.3.10

# ============ Retrieval ============
rank-bm25>=0.2.2

# ============ Environment ============
python-dotenv>=1.0.0
psutil>=5.9.0

# ============ API & Data ============
requests>=2.31.0
pandas>=2.0.0
tabulate>=0.9.0

# ============ Math ============
numpy>=1.24.0
